refactor: First pass to support 0.17.

+++ examples/common/lib.rs
@@ -116,7 +116,7 @@ impl VideoCapturePlugin {
                 };
                 let res = self.resolution;
                 // if let Some(res) = self.resolution {
-                window.resolution = [res.x, res.y].into();
+                window.resolution = [res.x as u32, res.y as u32].into();
                 // }
                 window
             }),
